// src/types/category.d.ts
export enum Category {
    ARTICULO_DE_OFICINA = "Articulo de Oficina",
    COMPUTACION = "Computacion",
    ELECTRONICA = "Electronica",
    FERRETERIA = "Ferreteria",
    FUNGIBLE  = "Fungible",
    MATERIAL_DE_ASEO = "Material de Aseo",
    SUPERMERCADO = "Supermercado",
    MOBILIARIO = "Mobiliario",
    MOBILIARIO_DE_OFICINA = "Mobiliario de Oficina",
    COCINA= "Cocina",
    ARTICULO_ELECTRICO = "Art. eléctrico",
  }